+++ vng_api_common/client.py
@@ -1,44 +1,73 @@
 """
-Interface to get a zds_client object for a given URL.
+Interface to get a client object for a given URL.
 """
 
-from typing import Optional
+import logging
+from typing import Any, Optional
 
-from django.apps import apps
 from django.conf import settings
 from django.utils.module_loading import import_string
 
-from zds_client import Client
+from ape_pie import APIClient
+from requests import JSONDecodeError, RequestException, Response
+
+logger = logging.getLogger(__name__)
 
 
-def get_client(url: str, url_is_api_root=False) -> Optional[Client]:
+class ClientError(RuntimeError):
+    pass
+
+
+# TODO: use more approriate method name
+def to_internal_data(response: Response) -> dict | list | None:
+    try:
+        response_json = response.json()
+    except JSONDecodeError:
+        logger.exception("Unable to parse json from response")
+        response_json = None
+
+    try:
+        response.raise_for_status()
+    except RequestException as exc:
+        if response.status_code >= 500:
+            raise
+        raise ClientError(response_json if response_json is not None else {}) from exc
+
+    assert response_json
+    return response_json
+
+
+class Client(APIClient):
+    def request(
+        self, method: str | bytes, url: str | bytes, *args, **kwargs
+    ) -> Response:
+
+        headers = kwargs.pop("headers", {})
+        headers.setdefault("Accept", "application/json")
+        headers.setdefault("Content-Type", "application/json")
+
+        kwargs["headers"] = headers
+
+        data = kwargs.get("data", {})
+
+        if data:
+            kwargs["json"] = data
+
+        return super().request(method, url, *args, **kwargs)
+
+
+def get_client(url: str) -> Client | None:
     """
     Get a client instance for the given URL.
-
-    If the setting CUSTOM_CLIENT_FETCHER is defined, then this callable is invoked.
-    Otherwise we fall back on the default implementation.
-
     If no suitable client is found, ``None`` is returned.
     """
-    custom_client_fetcher = getattr(settings, "CUSTOM_CLIENT_FETCHER", None)
-    if custom_client_fetcher:
-        client_getter = import_string(custom_client_fetcher)
-        return client_getter(url)
+    from zgw_consumers.client import build_client
+    from zgw_consumers.models import Service
 
-    # default implementation
-    Client = import_string(settings.ZDS_CLIENT_CLASS)
+    service: Optional[Service] = Service.get_service(url)
 
-    if url_is_api_root and not url.endswith("/"):
-        url = f"{url}/"
-
-    client = Client.from_url(url)
-    if client is None:
+    if not service:
+        logger.warning(f"No service configured for {url}")
         return None
 
-    APICredential = apps.get_model("vng_api_common", "APICredential")
-
-    if url_is_api_root:
-        client.base_url = url
-
-    client.auth = APICredential.get_auth(url)
-    return client
+    return build_client(service, client_factory=Client)