+++ vng_api_common/mocks.py
@@ -1,7 +1,10 @@
 import re
 from urllib.parse import urlparse
 
-from zds_client.client import UUID_PATTERN
+UUID_PATTERN = re.compile(
+    r"[0-9a-f]{8}\-[0-9a-f]{4}\-4[0-9a-f]{3}\-[89ab][0-9a-f]{3}\-[0-9a-f]{12}",
+    flags=re.I,
+)
 
 
 class Response: