+++ vng_api_common/notifications/constants.py
@@ -1,3 +1,0 @@
-# Exernally defined scopes.
-SCOPE_NOTIFICATIES_CONSUMEREN_LABEL = "notificaties.consumeren"
-SCOPE_NOTIFICATIES_PUBLICEREN_LABEL = "notificaties.publiceren"